+++ contracts/core/BotAuth.sol
@@ -6,13 +6,12 @@ import "../auth/AdminAuth.sol";
 
 /// @title Handles authorization of who can call the execution of strategies
 contract BotAuth is AdminAuth {
-
-    mapping (address => bool) public approvedCallers;
+    mapping(address => bool) public approvedCallers;
 
     /// @notice Checks if the caller is approved for the specific strategy
     /// @dev Currently auth callers are approved for all strategies
     /// @param _caller Address of the caller
-    function isApproved(uint, address _caller) public view returns (bool) {
+    function isApproved(uint256, address _caller) public view returns (bool) {
         return approvedCallers[_caller];
     }
 
@@ -27,5 +26,4 @@ contract BotAuth is AdminAuth {
     function removeCaller(address _caller) public onlyOwner {
         approvedCallers[_caller] = false;
     }
-
 }
