+++ contracts/core/DFSRegistry.sol
@@ -7,19 +7,20 @@ import "../utils/DefisaverLogger.sol";
 
 /// @title Stores all the important DFS addresses and can be changed (timelock)
 contract DFSRegistry is AdminAuth {
-
-    DefisaverLogger public constant logger = DefisaverLogger(0x5c55B921f590a89C1Ebe84dF170E655a82b62126);
+    DefisaverLogger public constant logger = DefisaverLogger(
+        0x5c55B921f590a89C1Ebe84dF170E655a82b62126
+    );
 
     struct Entry {
         address contractAddr;
-        uint waitPeriod;
-        uint changeStartTime;
+        uint256 waitPeriod;
+        uint256 changeStartTime;
         bool inChange;
         bool exists;
     }
 
-    mapping (bytes32 => Entry) public entries;
-    mapping (bytes32 => address) public pendingAddresses;
+    mapping(bytes32 => Entry) public entries;
+    mapping(bytes32 => address) public pendingAddresses;
 
     /// @notice Given an contract id returns the registred address
     /// @dev Id is kecceak256 of the contract name
@@ -39,7 +40,11 @@ contract DFSRegistry is AdminAuth {
     /// @param _id Id of contract
     /// @param _contractAddr Address of the contract
     /// @param _waitPeriod Amount of time to wait before a contract address can be changed
-    function addNewContract(bytes32 _id, address _contractAddr, uint _waitPeriod) public onlyOwner {
+    function addNewContract(
+        bytes32 _id,
+        address _contractAddr,
+        uint256 _waitPeriod
+    ) public onlyOwner {
         require(!entries[_id].exists, "Entry id already exists");
 
         entries[_id] = Entry({
@@ -50,7 +55,12 @@ contract DFSRegistry is AdminAuth {
             exists: true
         });
 
-        logger.Log(address(this), msg.sender, "AddNewContract", abi.encode(_id, _contractAddr, _waitPeriod));
+        logger.Log(
+            address(this),
+            msg.sender,
+            "AddNewContract",
+            abi.encode(_id, _contractAddr, _waitPeriod)
+        );
     }
 
     /// @notice Starts an address change for an existing entry
@@ -65,7 +75,12 @@ contract DFSRegistry is AdminAuth {
 
         pendingAddresses[_id] = _newContractAddr;
 
-        logger.Log(address(this), msg.sender, "StartChange", abi.encode(_id, entries[_id].contractAddr, _newContractAddr));
+        logger.Log(
+            address(this),
+            msg.sender,
+            "StartChange",
+            abi.encode(_id, entries[_id].contractAddr, _newContractAddr)
+        );
     }
 
     /// @notice Changes new contract address, correct time must have passed
@@ -74,7 +89,10 @@ contract DFSRegistry is AdminAuth {
     function approveContractChange(bytes32 _id) public onlyOwner {
         require(entries[_id].exists, "Entry id doesn't exists");
         require(entries[_id].inChange, "Entry not in change process");
-        require((entries[_id].changeStartTime + entries[_id].waitPeriod) > block.timestamp, "Change not ready yet"); // solhint-disable-line
+        require(
+            (entries[_id].changeStartTime + entries[_id].waitPeriod) > block.timestamp,
+            "Change not ready yet"
+        ); // solhint-disable-line
 
         address oldContractAddr = entries[_id].contractAddr;
         entries[_id].contractAddr = pendingAddresses[_id];
@@ -83,7 +101,12 @@ contract DFSRegistry is AdminAuth {
 
         pendingAddresses[_id] = address(0);
 
-        logger.Log(address(this), msg.sender, "ApproveChange", abi.encode(_id, oldContractAddr, entries[_id].contractAddr));
+        logger.Log(
+            address(this),
+            msg.sender,
+            "ApproveChange",
+            abi.encode(_id, oldContractAddr, entries[_id].contractAddr)
+        );
     }
 
     /// @notice Cancel pending change
@@ -98,13 +121,18 @@ contract DFSRegistry is AdminAuth {
         entries[_id].inChange = false;
         entries[_id].changeStartTime = 0;
 
-        logger.Log(address(this), msg.sender, "CancelChange", abi.encode(_id, oldContractAddr, entries[_id].contractAddr));
+        logger.Log(
+            address(this),
+            msg.sender,
+            "CancelChange",
+            abi.encode(_id, oldContractAddr, entries[_id].contractAddr)
+        );
     }
 
     /// @notice Changes wait period for an entry
     /// @param _id Id of contract
     /// @param _newWaitPeriod New wait time, must be bigger than before
-    function changeWaitPeriod(bytes32 _id, uint _newWaitPeriod) public onlyOwner {
+    function changeWaitPeriod(bytes32 _id, uint256 _newWaitPeriod) public onlyOwner {
         require(entries[_id].exists, "Entry id doesn't exists");
         require(_newWaitPeriod > entries[_id].waitPeriod, "New wait period must be bigger");
 
@@ -112,5 +140,4 @@ contract DFSRegistry is AdminAuth {
 
         logger.Log(address(this), msg.sender, "ChangeWaitPeriod", abi.encode(_id, _newWaitPeriod));
     }
-
 }
