+++ contracts/core/StrategyData.sol
@@ -5,7 +5,6 @@ pragma experimental ABIEncoderV2;
 
 /// @title Struct Data in a separate contract soit can be used in multiple places
 contract StrategyData {
-
     struct Trigger {
         bytes32 id;
         bytes data;
@@ -20,7 +19,7 @@ contract StrategyData {
         string name;
         address proxy;
         bool active;
-        uint[] triggerIds;
-        uint[] actionIds;
+        uint256[] triggerIds;
+        uint256[] actionIds;
     }
 }
