+++ contracts/core/SubscriptionProxy.sol
@@ -9,10 +9,15 @@ import "./Subscriptions.sol";
 
 /// @title Handles auth and calls subscription contract
 contract SubscriptionProxy is StrategyData {
-
     address public constant FACTORY_ADDRESS = 0x5a15566417e6C1c9546523066500bDDBc53F88C7;
 
-    function subscribe(string memory _name, address _executorAddr, address _subAddr, Trigger[] memory _triggers, Action[] memory _actions) public {
+    function subscribe(
+        string memory _name,
+        address _executorAddr,
+        address _subAddr,
+        Trigger[] memory _triggers,
+        Action[] memory _actions
+    ) public {
         address currAuthority = address(DSAuth(address(this)).authority());
         DSGuard guard = DSGuard(currAuthority);
 
@@ -26,12 +31,17 @@ contract SubscriptionProxy is StrategyData {
         Subscriptions(_subAddr).subscribe(_name, _triggers, _actions);
     }
 
-    function update(address _subAddr, uint _subId, Trigger[] memory _triggers, Action[] memory _actions) public {
+    function update(
+        address _subAddr,
+        uint256 _subId,
+        Trigger[] memory _triggers,
+        Action[] memory _actions
+    ) public {
         Subscriptions(_subAddr).update(_subId, _triggers, _actions);
     }
 
     // TODO: should we remove permission if no more strategies left?
-    function unsubscribe(address _subAddr, uint _subId) public {
+    function unsubscribe(address _subAddr, uint256 _subId) public {
         Subscriptions(_subAddr).unsubscribe(_subId);
     }
 }
