+++ contracts/core/Subscriptions.sol
@@ -9,8 +9,9 @@ import "./StrategyData.sol";
 
 /// @title Storage of actions and triggers which can be added/removed and modified
 contract Subscriptions is StrategyData {
-
-    DefisaverLogger public constant logger = DefisaverLogger(0x5c55B921f590a89C1Ebe84dF170E655a82b62126);
+    DefisaverLogger public constant logger = DefisaverLogger(
+        0x5c55B921f590a89C1Ebe84dF170E655a82b62126
+    );
 
     Strategy[] internal strategies;
     Action[] internal actions;
@@ -19,37 +20,37 @@ contract Subscriptions is StrategyData {
     /// @notice Subscribes a new strategy for a user
     /// @param _triggers Array of trigger data
     /// @param _actions Array of action data
-    function subscribe(string memory _name, Trigger[] memory _triggers, Action[] memory _actions) public {
-        uint[] memory triggerIds = new uint[](_triggers.length);
-        uint[] memory actionsIds = new uint[](_actions.length);
+    function subscribe(
+        string memory _name,
+        Trigger[] memory _triggers,
+        Action[] memory _actions
+    ) public {
+        uint256[] memory triggerIds = new uint256[](_triggers.length);
+        uint256[] memory actionsIds = new uint256[](_actions.length);
 
         // Populate triggers
-        for (uint i = 0; i < _triggers.length; ++i) {
-            triggers.push(Trigger({
-                id: _triggers[i].id,
-                data: _triggers[i].data
-            }));
+        for (uint256 i = 0; i < _triggers.length; ++i) {
+            triggers.push(Trigger({id: _triggers[i].id, data: _triggers[i].data}));
 
             triggerIds[i] = triggers.length - 1;
         }
 
         // Populate actions
-        for (uint i = 0; i < _actions.length; ++i) {
-            actions.push(Action({
-                id: _actions[i].id,
-                data: _actions[i].data
-            }));
+        for (uint256 i = 0; i < _actions.length; ++i) {
+            actions.push(Action({id: _actions[i].id, data: _actions[i].data}));
 
             actionsIds[i] = actions.length - 1;
         }
 
-        strategies.push(Strategy({
-            name: _name,
-            proxy: msg.sender,
-            active: true,
-            triggerIds: triggerIds,
-            actionIds: actionsIds
-        }));
+        strategies.push(
+            Strategy({
+                name: _name,
+                proxy: msg.sender,
+                active: true,
+                triggerIds: triggerIds,
+                actionIds: actionsIds
+            })
+        );
 
         logger.Log(address(this), msg.sender, "Subscribe", abi.encode(strategies.length - 1));
     }
@@ -60,25 +61,23 @@ contract Subscriptions is StrategyData {
     /// @param _subId Subscription id
     /// @param _triggers Array of trigger data
     /// @param _actions Array of action data
-    function update(uint _subId, Trigger[] memory _triggers, Action[] memory _actions) public {
+    function update(
+        uint256 _subId,
+        Trigger[] memory _triggers,
+        Action[] memory _actions
+    ) public {
         Strategy memory s = strategies[_subId];
         require(s.proxy != address(0), "Strategy does not exist");
         require(msg.sender == s.proxy, "Proxy not strategy owner");
 
         // update triggers
-        for (uint i = 0; i < _triggers.length; ++i) {
-            triggers[s.triggerIds[i]] = Trigger({
-                id: _triggers[i].id,
-                data: _triggers[i].data
-            });
+        for (uint256 i = 0; i < _triggers.length; ++i) {
+            triggers[s.triggerIds[i]] = Trigger({id: _triggers[i].id, data: _triggers[i].data});
         }
 
         // update actions
-        for (uint i = 0; i < _actions.length; ++i) {
-            actions[s.actionIds[i]] = Action({
-                id: _actions[i].id,
-                data: _actions[i].data
-            });
+        for (uint256 i = 0; i < _actions.length; ++i) {
+            actions[s.actionIds[i]] = Action({id: _actions[i].id, data: _actions[i].data});
         }
 
         logger.Log(address(this), msg.sender, "Update", abi.encode(_subId));
@@ -86,7 +85,7 @@ contract Subscriptions is StrategyData {
 
     /// @notice Unsubscribe an existing strategy
     /// @param _subId Subscription id
-    function unsubscribe(uint _subId) public {
+    function unsubscribe(uint256 _subId) public {
         Strategy memory s = strategies[_subId];
         require(s.proxy != address(0), "Strategy does not exist");
         require(msg.sender == s.proxy, "Proxy not strategy owner");
@@ -96,7 +95,6 @@ contract Subscriptions is StrategyData {
         logger.Log(address(this), msg.sender, "Unsubscribe", abi.encode(_subId));
     }
 
-
     function getProxyOwner(address _proxy) internal returns (address proxyOwner) {
         proxyOwner = IDSProxy(_proxy).owner();
         require(proxyOwner != address(0), "No proxy");
@@ -104,20 +102,19 @@ contract Subscriptions is StrategyData {
 
     ///////////////////// VIEW ONLY FUNCTIONS ////////////////////////////
 
-    function getTrigger(uint _triggerId) public view returns (Trigger memory) {
+    function getTrigger(uint256 _triggerId) public view returns (Trigger memory) {
         return triggers[_triggerId];
     }
 
-    function getAction(uint _actionId) public view returns (Action memory) {
+    function getAction(uint256 _actionId) public view returns (Action memory) {
         return actions[_actionId];
     }
 
-    function getStreategyCount() public view returns (uint) {
+    function getStreategyCount() public view returns (uint256) {
         return strategies.length;
     }
 
-    function getStrategy(uint _subId) public view returns (Strategy memory) {
+    function getStrategy(uint256 _subId) public view returns (Strategy memory) {
         return strategies[_subId];
     }
-
 }
